python-pathway/requirements.txt
yfinance==0.2.40
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
ta==0.11.0
langchain-community==0.0.32
# langchain-core removed because langchain-community will install the compatible version
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
asyncio-throttle


